add compatibility test for various Gradle versions
